backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.17.0
python-dotenv==0.21.0   
mysql-connector-python
